evaluation_scripts/requirements.txt
# Training
scipy==1.9.3
opencv-python==4.7.0.68
tensorflow==2.11.0
# Executing
matplotlib==3.6.1
# Hendrycks
scikit-image==0.20.0
Wand==0.6.11
Pillow==9.5.0
# Evaluation
scikit-learn==1.3.0
